refactor(database): split adapters into adapters.ts and re-export from index
